Pre-emptively commit if limits will be exceeded
We can use the measured increase in memory usage of the previous iteration of
a StreamThread to predict if we're likely to exceed our limit in the
_next_ iteration. If we will, we pre-emptively commit on the current
iteration, instead of allowing the threshold to be crossed on the next
iteration.

This should prevent most OOMs as a result of crossing our threshold in
the middle of the next iteration. Although it's still possible to go OOM
if the first iteration attempts to write too much, or if write patterns
are highly varied.