Fix race with StateUpdater on init of new stores
When initializing a new store, its current offset is unknown. If the
StateUpdater triggers a `commit` of this store before some records have
been written/restored to it, the offset will still be unknown.

In this case it's important that we write a sentinel "unknown" value to
the store, so that if the store is subsequently closed and re-opened,
it's not mistaken for a non-empty store without an offset and wiped.

To do this, we use the `OFFSET_UNKOWN` value, which was used in the old
`.checkpoint` files for the same purpose.