Don't throw TaskCorruptedException on Timeout
Previously, under EOS, any timeout in communication with Kafka would
cause a `TaskCorruptedException` to be thrown, which would wipe the
stores for that `Task`.

This was because records are proactively written to a `StateStore`
before the Kafka transaction has been committed, so any timeout of this
Kafka transaction would mean that some records were successfully written
to the local store that failed to be written to its changelog, violating
consistency.

Now that our `StateStore` API is transactional, we can guarantee that
records written to them won't be persisted to disk until after the Kafka
transaction successfully commits.

This means that a timeout simply means we need to close-and-revive the
`Task`, which dumps the current `StateStore` transaction without
committing its records to the local disk.

We will now only throw a `TaskCorruptedException`, and hence wipe local
stores, under two conditions:

  1. The `.checkpoint` file is missing or corrupt, which means we can't
     reason about the state of the local stores, so we must wipe them
     and rebuild; or,
  2. The local offset for a store is no longer available in the
     changelog in Kafka, which indicates the changelog was wiped/reset,
     so we should do the same to the local store.