Only open RocksDB once during init
Parsing an error message to detect missing column families is pretty
nasty, and it involved us potentially having to open the database *3*
times.

It's _possible_ (though not proven) that this was leading to the Block
Cache being doubled under certain circumstances.

Refactoring this to proactively determine the available column families
and creating them allows us to open RocksDB only once.

This should not affect performance, as stores are rarely initialized,
and these operations should be no more expensive than the previous
strategy.