Ensure StateStores are only wiped when corrupt
When a `TaskCorruptedException` is handled, the
`ProcessorStateManager.changelogPartition` for the store is marked as
"corrupted".

We now check this when closing "dirty", and only wipe the state stores
for the Task when at least one changelog is marked as dirty.

This ensures that the Task is being closed due to a
`TaskCorruptedException`, and not some other exception (e.g.
`KafkaException`, `TimeoutException`, Producer fencing, etc.).

Some tests were previously relying incorrectly on explicitly flushing
the `StateStore` in their `Processor`, which broke their assertions now
that we're no longer wiping stores on a non-corrupt error.