Improve state restore performance
Calling `writeOffset` for every record during a restore is expensive,
because it requires:

  a. serializing the `TopicPartition`
  b. adding a write operation to the RocksDB write batch, which has to
     process each one into the memtable on commit/write

Instead, we track the latest offset(s) as we process a batch, and only
`writeOffset` once (for each changelog partition) at the end of the
batch.

We also add some caching to `writeOffset`, so that we don't need to
re-serialize the `TopicPartition` on each commit; instead, it will cost
a single `HashMap` look-up, to a very small cardinality (usually just 1)
hashtable.