Change StateStore#flush default to no-op
Also clarifies the JavaDoc for `StateStore#flush` and
`StateStore#commit`.

All internal uses of `flush` have been converted to `commit`, where
appropriate. `flush` is retained for backwards compatibility with
user-code that explicitly flushes the `StateStore`.

Nearly all of this code will be custom user `Processor`s that call
`StateStore#flush()` in an attempt to persist their results.

Now that we have strict consistency guarantees both provided by and
expected of StateStores, we cannot just flush them to disk when the user
requests it - we need to flush only in response to a successful changelog
transaction commit.

Switching this to a no-op prevents existing user-code from violating
these guarantees.

The JavaDoc for both `flush` and `commit` has been updated to discourage
users from attempting to use these methods, and points them towards the
correct way to flush `StateStores`: `ProcessingContext#commit()`.