Add commit rate/latency metrics
We now have state-level metrics for calls to `StateStore#commit` that is
distinct from the metrics for `StateStore#flush`.

The `flush` metrics _should_ never report any value, but we continue to
register them for backwards compatibility; they are effectively
deprecated.

The only case where `flush` metrics will provide a value is if users are
explicitly calling `StateStore#flush()` from a custom `Processor`, even
though this method is effectively a no-op.