KIP-990: SUSPEND Tasks on DeserializationException
Adds a new `SUSPEND` option to `DeserializationHandlerResponse`, which
suspends the current `Task` while continuing to process other tasks.

When suspending, the offset(s) of the last successfully processed
record(s) will be committed for the `Task`, to ensure that previously
processed records will not be reprocessed when the Task is resumed.