Split open-iterators->iterators-(created|deleted)
The `open-iterators` metric was usually reporting zero, despite lots of
Iterators apparently closing according to `iterators-duration`.

This is likely because the Iterator duration is so short that most of
the time when the metrics are queried, no Iterators are currently open.

Instead, we now have two metrics:

- `iterators-created-total`: measures the number of Iterator instances that
  have been created.
- `iterators-deleted-total`: measures the number of Iterator instances that
  have been deleted/closed.

Users can then calculate the instantaneous number of Iterators open at
any given moment by subtracting deleted from created.