Disable Transactional stores when READ_UNCOMMITTED
If the `processing.mode` is `exactly_once(_v2|_beta)`, _and_ the
`default.state.isolation.level` is `READ_UNCOMMITTED`, we write directly
to stores without buffering writes, and wipe local state on-error, as
before KIP-892.

This combination must be selected explicitly by users, because the
default isolation level is set to `READ_COMMITTED` when exactly-once
streaming is enabled.