KAFKA-14412: Transactional StateStores
See [KIP-892: Transactional Semantics for StateStores][1] for a detailed
explanation of the design.

---

Writes to StateStores are now isolated from the underlying RocksDB
database unti the changelog transaction has been committed. We do this
by writing first to a RocksDB `WriteBatchWithIndex`, which on-commit
will be written to the RocksDB database.

Now that writes to a StateStore are isolated by a "transaction buffer"
(`WriteBatchWithIndex`), we can also isolate them from interactive query
threads. We introduce a new configuration,
`default.state.isolation.level`, which defaults to `READ_UNCOMMITTED` to
provide the existing behaviour.

When `processing.mode` is set to `exactly-once(-v2|-beta)`, the default
swtiches to `READ_COMMITTED`, as this will be the desired behaviour for
the vast majority of use-cases.

Crucially, this ensures that in the event of an error, state stores no
longer need to be wiped under EOS, unless the user has explicitly
selected the `READ_UNCOMMITTED` isolation level.

---

To ensure consistency, we now store offsets in a secondary column-family
in RocksDB (`__offsets`). These offsets are updated on-commit. We
guarantee that the `__offsets` memtables and data memtables are flushed
atomically, together, using RocksDB's Atomic Flush. This allows us to
remove the explicit `flush()` calls that previously forcibly flushed
memtables to disk. RocksDB now entirely controls memtable flush behaviour
via its configuration.

We continue writing offsets to the `.checkpoint` file, to ensure that
during rebalance, we can send offsets for currently closed stores,
without having to open them just to read their offset. The `.checkpoint`
file is only updated on store close for this reason. If the store is
open, the currently in-memory offset is used instead.

We also keep the `.checkpoint` file for custom `StateStore`
implementations that do not manage their own offsets. For these stores,
the legacy behaviour will still be used.

---

The performance should be at least as good, if not better, than before,
as batched writes to RocksDB avoid locking the memtable for every write.

1: https://cwiki.apache.org/confluence/display/KAFKA/KIP-892%3A+Transactional+Semantics+for+StateStores