KAFKA-14412: Introduce KIP-1035 StateStore APIs
This adds the new StateStore APIs, introduced in KIP-1035.

Since `StateStore#flush()` is now deprecated in favor of
`StateStore#commit(Map)`, we replace all implementations and all calls
of this method with `StateStore#commit(Map)`.

Currently, there is no change of behaviour: `StateStore#commit(Map)`
is simply a rename of `StateStore#flush()`, that now accepts a `Map` of
`changelogOffsets`. However, this `Map` is not _yet_ used.

For this reason, `ProcessorStateManager` and `GlobalStateManagerImpl`
(the only two places to call `flush` that aren't either tests or
forwarding wrappers) provide an empty `Map` of offsets.

In a future commit, we will change the behaviour of
`ProcessorStateManager` and `GlobalStateManagerImpl` to provide a real
`Map` of changelog offsets, and we will change `RocksDBStore` to make
use of them.

The goal of this commit is to reduce the review burden of later PRs that
will actually change behaviour, while maintaining behaviour and
compatibility of trunk.

Some tests have had their `flush()` calls removed, because they were
unneecessary. Others have had their `flush()` call replaced with
`flushCache()`, because they were actually testing caching behaviour.
The remaining tests actually test or depend upon the flush/commit
behaviour of the store, so the `flush()` call has been replaced with
`commit(Collections.emptyMap())`.